requirements.txt
langchain
openai
streamlit
torch
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
beautifulsoup4
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
langchain_community
langchain-chroma
tabula-py 
pandas
pyPDF2
pdfplumber
duckduckgo-search
arxiv
Wikipedia
langchainhub
gradio
langchain_nvidia_ai_endpoints
rank_bm25
